+++ CHANGELOG.rst
@@ -77,6 +77,7 @@ Changed
 -------
 
 - Backend defaults values are no longer required (#2101).
+- QuantumCircuit properties more self-consistent and no longer need DAG (#1993).
 - The most connected subset in DenseLayout is now reduced bandwidth (#2021).
 - plot_histogram now allows sorting by Hamming distance from target_string (#2064).
 - FunctionalPulse is no longer a class and instead is a decorator, `functional_pulse`